requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
CacheControl==0.12.6
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1
django-crispy-forms==1.9.2
django-jazzmin==2.2.7
django-paypal==1.0.0
django-singletons==0.1.6
django-summernote==0.8.11.6
djsingleton==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isodate==0.6.0
isort==5.4.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
pylint==2.6.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.12.1
zeep==3.4.0
